requirements.txt
# Core AI/ML Dependencies
openai>=1.12.0
chromadb>=0.4.22
networkx>=3.2.1
datasets>=2.16.1

# Data Processing
pandas>=2.1.4
numpy>=1.24.3

# Visualization
matplotlib>=3.8.2
plotly>=5.17.0
pyvis>=0.3.2

# Configuration and Environment
python-dotenv>=1.0.0

# Natural Language Processing
spacy>=3.7.2

# Web Interface
streamlit>=1.31.1

# Development and Testing
jupyter>=1.0.0
scikit-learn>=1.4.0

# Utilities
tqdm>=4.66.1
click>=8.1.7